fix(#417): sharpen lib/ vs contracts/ boundary — move implementation files to correct tiers
- Document tier-neutral layer rules in KERNEL-ARCHITECTURE.md §3.1
- Move db_base.py, validators.py from contracts/ → lib/ (implementation helpers, not pure types)
- Move filters.py from core/ → fuse/ (single-layer code stays in its layer)
- Replace core.glob_fast dependency with stdlib fnmatch in filters.py
- Fix `any` (builtin) → `Any` (typing) type annotation in filters.py
- Remove type:ignore[no-any-return] from filters.py Rust FFI call
- Update all 12 callers across bricks, connectors, contracts, fuse, storage, tests
