+++ src/nexus/fuse/ops/metadata_handler.py
@@ -9,8 +9,8 @@ from typing import TYPE_CHECKING, Any, cast
 
 from fuse import FuseOSError
 
-from nexus.core.filters import is_os_metadata_file
 from nexus.core.virtual_views import should_add_virtual_views
+from nexus.fuse.filters import is_os_metadata_file
 from nexus.fuse.ops._shared import (
     FUSESharedContext,
     build_dir_attrs,
