+++ src/nexus/fuse/ops/mutation_handler.py
@@ -9,7 +9,7 @@ from typing import Any
 
 from fuse import FuseOSError
 
-from nexus.core.filters import is_os_metadata_file
+from nexus.fuse.filters import is_os_metadata_file
 from nexus.fuse.ops._shared import (
     FUSESharedContext,
     check_namespace_visible,
